Default impl for EthApiSpec
