+++ crates/rpc/rpc-eth-api/Cargo.toml
@@ -29,6 +29,7 @@ reth-chainspec.workspace = true
 reth-execution-types.workspace = true
 reth-rpc-eth-types.workspace = true
 reth-rpc-server-types.workspace = true
+reth-network-api.workspace = true
 
 # ethereum
 alloy-dyn-abi = { workspace = true, features = ["eip712"] }
