+++ crates/rpc/rpc/src/eth/helpers/spec.rs
@@ -1,13 +1,7 @@
-use std::sync::Arc;
-
-use reth_chainspec::{ChainInfo, ChainSpec};
-use reth_errors::{RethError, RethResult};
-use reth_evm::ConfigureEvm;
 use reth_network_api::NetworkInfo;
-use reth_primitives::{Address, U256, U64};
-use reth_provider::{BlockReaderIdExt, ChainSpecProvider, EvmEnvProvider, StateProviderFactory};
+use reth_primitives::U256;
+use reth_provider::{BlockNumReader, ChainSpecProvider};
 use reth_rpc_eth_api::helpers::EthApiSpec;
-use reth_rpc_types::{SyncInfo, SyncStatus};
 use reth_transaction_pool::TransactionPool;
 
 use crate::EthApi;
@@ -15,57 +9,23 @@ use crate::EthApi;
 impl<Provider, Pool, Network, EvmConfig> EthApiSpec for EthApi<Provider, Pool, Network, EvmConfig>
 where
     Pool: TransactionPool + 'static,
-    Provider:
-        BlockReaderIdExt + ChainSpecProvider + StateProviderFactory + EvmEnvProvider + 'static,
+    Provider: BlockNumReader + ChainSpecProvider + 'static,
     Network: NetworkInfo + 'static,
-    EvmConfig: ConfigureEvm,
+    EvmConfig: Send + Sync,
 {
-    /// Returns the current ethereum protocol version.
-    ///
-    /// Note: This returns an [`U64`], since this should return as hex string.
-    async fn protocol_version(&self) -> RethResult<U64> {
-        let status = self.network().network_status().await.map_err(RethError::other)?;
-        Ok(U64::from(status.protocol_version))
+    fn provider(&self) -> impl ChainSpecProvider + BlockNumReader {
+        self.inner.provider()
     }
 
-    /// Returns the chain id
-    fn chain_id(&self) -> U64 {
-        U64::from(self.network().chain_id())
+    fn network(&self) -> impl NetworkInfo {
+        self.inner.network()
     }
 
-    /// Returns the current info for the chain
-    fn chain_info(&self) -> RethResult<ChainInfo> {
-        Ok(self.provider().chain_info()?)
+    fn starting_block(&self) -> U256 {
+        self.inner.starting_block()
     }
 
-    fn accounts(&self) -> Vec<Address> {
-        self.inner.signers().read().iter().flat_map(|s| s.accounts()).collect()
-    }
-
-    fn is_syncing(&self) -> bool {
-        self.network().is_syncing()
-    }
-
-    /// Returns the [`SyncStatus`] of the network
-    fn sync_status(&self) -> RethResult<SyncStatus> {
-        let status = if self.is_syncing() {
-            let current_block = U256::from(
-                self.provider().chain_info().map(|info| info.best_number).unwrap_or_default(),
-            );
-            SyncStatus::Info(SyncInfo {
-                starting_block: self.inner.starting_block(),
-                current_block,
-                highest_block: current_block,
-                warp_chunks_amount: None,
-                warp_chunks_processed: None,
-            })
-        } else {
-            SyncStatus::None
-        };
-        Ok(status)
-    }
-
-    fn chain_spec(&self) -> Arc<ChainSpec> {
-        self.inner.provider().chain_spec()
+    fn signers(&self) -> &parking_lot::RwLock<Vec<Box<dyn reth_rpc_eth_api::helpers::EthSigner>>> {
+        self.inner.signers()
     }
 }
